setup-data/requirements.txt
kafka-python==2.0.2
clickhouse-driver==0.2.6
psycopg2-binary==2.9.6
pyyaml==6.0